core-engine/requirements.txt
numpy<1.24
pandas
scikit-learn
MetaTrader5
vectorbt
requests
pyyaml
python-dotenv
feedparser
pydantic>=2.0
pytz
aiosqlite
